requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
httpx==0.27.0
cryptography==43.0.1
prometheus-client==0.20.0
google-cloud-firestore==2.16.0
google-cloud-kms==2.23.0
boto3==1.34.162
azure-keyvault-keys==4.9.0
azure-identity==1.17.1
python-json-logger==2.0.7
slowapi==0.1.9
google-auth==2.33.0
openai==1.51.2
requests==2.32.3
pytest==8.3.2
jinja2==3.1.4
python-multipart==0.0.9
mypy==1.11.1
